Update addLink behavior to handle editing link without a full selection.
When your cursor is within a node of linked text and the link dialog is
opened, the intention is almost always to edit the link for the full
node rather than insert a new link in place of the text for which the
cursor is sitting.

Here's how this interaction worked before:

https://share.cleanshot.com/1NYHrBxP

With this change, editing a link when the cursor is within a linked
node will edit the link for the entire linked text rather than replacing
or inserting the new link in place of or in the midst of the linked
node text.

Here is the change in action:

https://share.cleanshot.com/gLjDmjwN