Link handling tweaks (#230)
* Focus the editor when link dialog is closed.

This is especially useful when the ESC key is hit. Without this, the
link dialog disappears, you expect to be able to continue editing, but
the editor does not have focus. Now it does!

* Update addLink behavior to handle editing link without a full selection.

When your cursor is within a node of linked text and the link dialog is
opened, the intention is almost always to edit the link for the full
node rather than insert a new link in place of the text for which the
cursor is sitting.

Here's how this interaction worked before:

https://share.cleanshot.com/1NYHrBxP

With this change, editing a link when the cursor is within a linked
node will edit the link for the entire linked text rather than replacing
or inserting the new link in place of or in the midst of the linked
node text.

Here is the change in action:

https://share.cleanshot.com/gLjDmjwN

* Show invalid URL message when link validity is broken.

Prior to this the text field would turn red, but the "Not a valid URL"
message did not appear to the user.

* Create violet-queens-perform.md

---------

Co-authored-by: Konnor Rogers <konnor5456@gmail.com>