Syntax highlighting documentation updates.
Proposed for your consideration. Two things are addressed in here.

First is a function to strip all hljs-* spans from existing code when
editing with rhino. This allows CodeBlockLowlight to then properly re-
parse the code blocks for highlighting. Without this, the spans do
appear in the editor's highlighted code blocks, but some newlines are
missing. The missing newlines will be when they are at the end of a line
with only whitespace after them. Those newlines get lost in processing
if these spans are not first stripped.

Second is for DB storage, decoding the code content before we re-
process it for storage. Without decoding first, some HTML entities will
get double-encoded.

I also changed the #highlightCodeblocks function to no longer track
highlighting with a has-highlighted HTML attribute. In my testing this
was leading to the JS sometimes skipping highlighting upon DB storage.
I'm not sure why this tracking existed before, whether it was
inadvertent double-execution of this function or if it related to not
removing spans before reprocessing edited code, which we now do in
#stripHljsSpans. I don't _think_ it's necessary, but I could find out
I'm wrong.

I adapted this documentation from our Pika.page code. Apologies for any
typos/mistakes!